misc little change
git-svn-id: svn://23.27.117.66/h2/trunk@231 54bad739-14b1-1b4c-9595-0bb2987a2747

+++ lv/src/lv/FileSystem/RawFileReader.cpp
@@ -31,7 +31,7 @@ namespace lv
 			// read the whole file
 			infile.seekg(0, std::ios_base::beg);
 
-			init_buffer(*buf, len);
+			init_buffer(*buf, static_cast<size_t>(len));
 			infile.read(buffer::data(buf), len);
 		}
 		else
